requirements.txt
# Core dependencies
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0

# Image processing
Pillow==10.1.0

# AI/ML
openai==1.3.5

# Logging and monitoring
logfire==0.8.0

# Development dependencies
black==23.11.0
pylint==3.0.2 